@@ -21,21 +21,12 @@
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-1": "^6.13.0",
     "css-loader": "^0.24.0",
-    "enzyme": "^2.4.1",
-    "eslint": "^3.4.0",
-    "eslint-config-airbnb": "^10.0.1",
-    "eslint-import-resolver-webpack": "^0.5.1",
-    "eslint-loader": "^1.5.0",
-    "eslint-plugin-import": "^1.14.0",
-    "eslint-plugin-jsx-a11y": "^2.2.0",
-    "eslint-plugin-react": "^6.2.0",
     "expect": "^1.20.2",
     "fetch-mock": "^5.1.1",
     "jsdom": "^9.4.2",
     "less": "^2.7.1",
     "less-loader": "^2.2.3",
     "mocha": "^3.0.2",
-    "react-addons-perf": "^15.4.0",
     "react-addons-test-utils": "^15.4.0",
     "react-hot-loader": "^3.0.0-beta.6",
     "redux-devtools": "^3.3.1",
@@ -47,17 +38,18 @@
     "webpack-dev-server": "^1.16.2"
   },
   "dependencies": {
+    "@material-ui/core": "^3.2.2",
     "immutable": "^3.8.1",
     "material-ui": "^0.16.4",
     "moment": "^2.14.1",
-    "react": "^15.4.0",
+    "prop-types": "^15.6.2",
+    "react": "^16.5.2",
     "react-addons-pure-render-mixin": "^15.4.0",
-    "react-dom": "^15.4.0",
+    "react-dom": "^16.5.2",
     "react-redux": "^4.4.5",
-    "react-router": "^2.7.0",
+    "react-router": "^3.2.0",
     "react-router-redux": "^4.0.5",
     "react-table": "^6.8.6",
-    "react-tap-event-plugin": "^2.0.1",
     "redux": "^3.5.2",
     "reselect": "^2.5.3",
     "whatwg-fetch": "^1.0.0"
