@@ -1,4 +1,5 @@
-import React, { Component } from 'react';
+import PropTypes from "prop-types";
+import React, { Component } from "react";
 
 const isValid = input => input.match(/^((\d|[1-9]\d+)(\.\d{1,2})?|\.\d{1,2})$/);
 
@@ -25,12 +26,12 @@ class Fee extends Component {
 }
 
 Fee.propTypes = {
-  value: React.PropTypes.number.isRequired,
-  input: React.PropTypes.string.isRequired,
-  project: React.PropTypes.string.isRequired,
-  billedMinutes: React.PropTypes.number.isRequired,
-  onValueChange: React.PropTypes.func.isRequired,
-  onInputChange: React.PropTypes.func.isRequired,
+  value: PropTypes.number.isRequired,
+  input: PropTypes.string.isRequired,
+  project: PropTypes.string.isRequired,
+  billedMinutes: PropTypes.number.isRequired,
+  onValueChange: PropTypes.func.isRequired,
+  onInputChange: PropTypes.func.isRequired
 };
 
 export default Fee;
