@@ -1,4 +1,5 @@
-import React from 'react';
+import PropTypes from "prop-types";
+import React from "react";
 
 const BalanceViewBodyRow = (props) => (
   <th title={props.title}>
@@ -7,8 +8,8 @@ const BalanceViewBodyRow = (props) => (
 );
 
 BalanceViewBodyRow.propTypes = {
-  columnName: React.PropTypes.string.isRequired,
-  title: React.PropTypes.string.isRequired,
+  columnName: PropTypes.string.isRequired,
+  title: PropTypes.string.isRequired
 };
 
 export default BalanceViewBodyRow;
